Move `ColumnInfo` to dbt/artifacts
