+++ core/dbt/artifacts/resources/__init__.py
@@ -1,7 +1,12 @@
 from dbt.artifacts.resources.base import BaseResource, GraphResource
 
 # alias to latest resource definitions
-from dbt.artifacts.resources.v1.components import DependsOn, NodeVersion, RefArgs
+from dbt.artifacts.resources.v1.components import (
+    ColumnInfo,
+    DependsOn,
+    NodeVersion,
+    RefArgs,
+)
 from dbt.artifacts.resources.v1.documentation import Documentation
 from dbt.artifacts.resources.v1.exposure import (
     Exposure,
