+++ core/dbt/parser/common.py
@@ -1,4 +1,4 @@
-from dbt.artifacts.resources import NodeVersion
+from dbt.artifacts.resources import ColumnInfo, NodeVersion
 from dbt_common.contracts.constraints import ColumnLevelConstraint, ConstraintType
 from dbt.contracts.graph.unparsed import (
     HasColumnProps,
@@ -10,7 +10,7 @@ from dbt.contracts.graph.unparsed import (
     UnparsedModelUpdate,
 )
 from dbt.contracts.graph.unparsed import HasColumnTests, HasColumnDocs
-from dbt.contracts.graph.nodes import UnpatchedSourceDefinition, ColumnInfo
+from dbt.contracts.graph.nodes import UnpatchedSourceDefinition
 from dbt.parser.search import FileBlock
 from typing import List, Dict, Any, TypeVar, Generic, Union, Optional
 from dataclasses import dataclass
